improve package installation
change the way to check if a package is installed, by checking the command instead of relying on the package manager.
This makes this process cross-platform.

use common data definition for linux packages.
add asdf and rust installer. rust is used as a fallback if the package isn't installed with the package manager.
This allows to install rust packages on old distrib (debian).